Fix for an issue where the browser might hang.
The logic surrounding the decorated callback function was wrong - the
window.requestAnimationFrame system isn't an interval, it behaves as a
callback - so using setInterval on each frame would result in general
performance disasters if the system doesn't support a native
requestAnimationFrame system. Setting a timeout is much better, and
fixes this scenario.